Registration should get its own header
